+++ packages/s2-core/src/common/constant/options.ts
@@ -28,7 +28,9 @@ export enum LayoutWidthType {
 
 export const SPLIT_LINE_WIDTH = 1;
 
-export const DEFAULT_TREE_ROW_CELL_WIDTH = 120;
+export const DEFAULT_ROW_CELL_TREE_WIDTH = 120;
+export const DEFAULT_CELL_WIDTH = 96;
+export const DEFAULT_CELL_HEIGHT = 30;
 
 export const DEFAULT_CELL_TEXT_WORD_WRAP_STYLE: CellTextWordWrapStyle = {
   wordWrap: true,
@@ -46,12 +48,12 @@ export const DEFAULT_STYLE: S2Style = {
   },
   colCell: {
     ...DEFAULT_CELL_TEXT_WORD_WRAP_STYLE,
-    height: 30,
+    height: DEFAULT_CELL_HEIGHT,
   },
   dataCell: {
     ...DEFAULT_CELL_TEXT_WORD_WRAP_STYLE,
-    width: 96,
-    height: 30,
+    width: DEFAULT_CELL_WIDTH,
+    height: DEFAULT_CELL_HEIGHT,
   },
 } as const;